Templatised RawReader on FrameView and removed get_dataset_id.
