+++ lib/stages/HFBRawReader.hpp
@@ -11,6 +11,7 @@
 #include "bufferContainer.hpp" // for bufferContainer
 #include "datasetManager.hpp"  // for dset_id_t
 #include "visUtil.hpp"         // for frameID
+#include "HFBFrameView.hpp"
 
 #include <stdint.h> // for uint32_t
 #include <string>   // for string
@@ -23,7 +24,7 @@
  * This class inherits from the RawReader base class and reads raw 21cm absorber data
  * @author James Willis
  */
-class HFBRawReader : public RawReader {
+class HFBRawReader : public RawReader<HFBFrameView> {
 
 public:
     /// default constructor
@@ -49,9 +50,6 @@ public:
 protected:
     // Create an empty frame
     void create_empty_frame(frameID frame_id) override;
-
-    // Get dataset ID
-    dset_id_t& get_dataset_id(frameID frame_id) override;
 
 private:
     // The metadata
